Split tests and add a fixme test to capture limitation
